requirements.txt
allure-pytest==2.10.0
allure-python-commons==2.10.0
attrs==22.1.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
flake8==5.0.4
idna==3.3
iniconfig==1.1.1
mccabe==0.7.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.1
pydantic==1.10.2
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.1.3
requests==2.28.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.3.0
typish==1.9.3
urllib3==1.26.12
